Use a fake ArtistService to test suspending functions
Mockito doesn't directly support the mocking of a suspended function, so you can't easily simulate a delay in the response.

I suspect there's a neater way of doing the faking, because it ends up implementing a bit of mocking-type behaviour with the checks for whether it should throw, delay or return or whatever combination of the three.